requirements.txt
pytest
pytest-bdd
playwright
python-dotenv
